fix bug

+++ src/main/kotlin/com/cj/ticketsys/entities/Recommend.kt
@@ -5,7 +5,7 @@ import java.util.*
 class Recommend : PropertyEntity() {
     var id:Int = 0
     var refId:Int = 0
-    var type:Int = 0
+    var type:Short = 0
     var createTime:Date = Date()
     var displayOrder:Int = 0
 }
